WIP: Add support for parsing GitLab remote urls
Make GitLabUrlParser a non-static class and split class into multiple files
